+++ mycodo/inputs/scd4x_circuitpython.py
@@ -36,7 +36,7 @@ INPUT_INFORMATION = {
     'input_manufacturer': 'Sensirion',
     'input_name': 'SCD-4x (40, 41)',
     'input_name_short': 'SCD-4x',
-    'input_library': 'Adafruit-CircuitPython-SCD4x',
+    'input_library': 'Adafruit_CircuitPython_SCD4x',
     'measurements_name': 'CO2/Temperature/Humidity',
     'measurements_dict': measurements_dict,
     'url_manufacturer': 'https://www.sensirion.com/en/environmental-sensors/carbon-dioxide-sensors/carbon-dioxide-sensor-scd4x/',